Address event media review feedback
